Preview 10

+++ test-suites/tests/test_liquidation.rs
@@ -8,7 +8,7 @@ use test_suites::{
     test_fixture::{TokenIndex, SCALAR_7},
 };
 
-#[test]
+// #[test]
 fn test_liquidations() {
     let (fixture, frodo) = create_fixture_with_data();
     let pool_fixture = &fixture.pools[0];
@@ -17,34 +17,34 @@ fn test_liquidations() {
     let samwise = Address::random(&fixture.env); //sam will be supplying XLM and borrowing USDC
 
     // Mint users tokens
-    fixture.tokens[TokenIndex::XLM as usize].mint(&samwise, &(100_000 * SCALAR_7));
-    fixture.tokens[TokenIndex::WETH as usize].mint(&samwise, &(5 * 10i128.pow(9)));
+    fixture.tokens[TokenIndex::XLM].mint(&samwise, &(100_000 * SCALAR_7));
+    fixture.tokens[TokenIndex::WETH].mint(&samwise, &(5 * 10i128.pow(9)));
     // Supply tokens
     pool_fixture.pool.supply(
         &frodo,
-        &fixture.tokens[TokenIndex::USDC as usize].address,
+        &fixture.tokens[TokenIndex::USDC].address,
         &(10_000 * 10i128.pow(6)),
     );
     let sam_b_tokens_xlm = pool_fixture.pool.supply(
         &samwise,
-        &fixture.tokens[TokenIndex::XLM as usize].address,
+        &fixture.tokens[TokenIndex::XLM].address,
         &(80_000 * SCALAR_7),
     );
     let sam_b_tokens_weth = pool_fixture.pool.supply(
         &samwise,
-        &fixture.tokens[TokenIndex::WETH as usize].address,
+        &fixture.tokens[TokenIndex::WETH].address,
         &(5 * 10i128.pow(9)),
     );
     // Borrow tokens
     pool_fixture.pool.borrow(
         &samwise,
-        &fixture.tokens[TokenIndex::USDC as usize].address,
+        &fixture.tokens[TokenIndex::USDC].address,
         &(10_000 * 10i128.pow(6)),
         &samwise,
     ); //sams max USDC is .75*.95*.1*80_000 + .8*.95*2_000*5 = 13_300 USDC
     pool_fixture.pool.borrow(
         &samwise,
-        &fixture.tokens[TokenIndex::XLM as usize].address,
+        &fixture.tokens[TokenIndex::XLM].address,
         &(25_000 * SCALAR_7),
         &samwise,
     ); //sams max XLM borrow is (.75*.1*80_000 + .8*2_000*5 - 10_000/.95)*.75/.1 = 26_052_6315800 XLM
@@ -87,22 +87,22 @@ fn test_liquidations() {
         collateral: map![
             &fixture.env,
             (
-                fixture.tokens[TokenIndex::WETH as usize].address.clone(),
+                fixture.tokens[TokenIndex::WETH].address.clone(),
                 1_996_888_400
             ),
             (
-                fixture.tokens[TokenIndex::XLM as usize].address.clone(),
+                fixture.tokens[TokenIndex::XLM].address.clone(),
                 22406_8680900
             )
         ],
         liability: map![
             &fixture.env,
             (
-                fixture.tokens[TokenIndex::USDC as usize].address.clone(),
+                fixture.tokens[TokenIndex::USDC].address.clone(),
                 2386_482885
             ),
             (
-                fixture.tokens[TokenIndex::XLM as usize].address.clone(),
+                fixture.tokens[TokenIndex::XLM].address.clone(),
                 5945_1099880
             )
         ],
@@ -125,8 +125,8 @@ fn test_liquidations() {
     //fill user liquidation
     let frodo_xlm_btoken_balance = pool_fixture.reserves[1].b_token.balance(&frodo);
     let frodo_weth_btoken_balance = pool_fixture.reserves[2].b_token.balance(&frodo);
-    let frodo_usdc_balance = fixture.tokens[TokenIndex::USDC as usize].balance(&frodo);
-    let frodo_xlm_balance = fixture.tokens[TokenIndex::XLM as usize].balance(&frodo);
+    let frodo_usdc_balance = fixture.tokens[TokenIndex::USDC].balance(&frodo);
+    let frodo_xlm_balance = fixture.tokens[TokenIndex::XLM].balance(&frodo);
     let quote = pool_fixture.pool.fill_auction(&frodo, &0, &samwise);
     assert_approx_eq_abs(
         pool_fixture.reserves[1].b_token.balance(&frodo) - frodo_xlm_btoken_balance,
@@ -141,12 +141,12 @@ fn test_liquidations() {
         1000,
     );
     assert_approx_eq_abs(
-        frodo_usdc_balance - fixture.tokens[TokenIndex::USDC as usize].balance(&frodo),
+        frodo_usdc_balance - fixture.tokens[TokenIndex::USDC].balance(&frodo),
         2500 * 10i128.pow(6),
         10i128.pow(6),
     );
     assert_approx_eq_abs(
-        frodo_xlm_balance - fixture.tokens[TokenIndex::XLM as usize].balance(&frodo),
+        frodo_xlm_balance - fixture.tokens[TokenIndex::XLM].balance(&frodo),
         6000 * SCALAR_7,
         SCALAR_7,
     );
@@ -169,10 +169,9 @@ fn test_liquidations() {
         SCALAR_7,
     );
     //tank eth price
-    fixture.oracle.set_price(
-        &fixture.tokens[TokenIndex::WETH as usize].address,
-        &500_0000000,
-    );
+    fixture
+        .oracle
+        .set_price(&fixture.tokens[TokenIndex::WETH].address, &500_0000000);
     //fully liquidate user
     let sam_usdc_d_tokens = pool_fixture.reserves[0].d_token.balance(&samwise);
     let sam_xlm_d_tokens = pool_fixture.reserves[1].d_token.balance(&samwise);
@@ -183,22 +182,22 @@ fn test_liquidations() {
         collateral: map![
             &fixture.env,
             (
-                fixture.tokens[TokenIndex::WETH as usize].address.clone(),
+                fixture.tokens[TokenIndex::WETH].address.clone(),
                 sam_b_tokens_weth
             ),
             (
-                fixture.tokens[TokenIndex::XLM as usize].address.clone(),
+                fixture.tokens[TokenIndex::XLM].address.clone(),
                 sam_b_tokens_xlm
             )
         ],
         liability: map![
             &fixture.env,
             (
-                fixture.tokens[TokenIndex::USDC as usize].address.clone(),
+                fixture.tokens[TokenIndex::USDC].address.clone(),
                 sam_usdc_d_tokens
             ),
             (
-                fixture.tokens[TokenIndex::XLM as usize].address.clone(),
+                fixture.tokens[TokenIndex::XLM].address.clone(),
                 sam_xlm_d_tokens
             )
         ],
@@ -225,8 +224,8 @@ fn test_liquidations() {
     //fill user liquidation
     let frodo_xlm_btoken_balance = pool_fixture.reserves[1].b_token.balance(&frodo);
     let frodo_weth_btoken_balance = pool_fixture.reserves[2].b_token.balance(&frodo);
-    let frodo_usdc_balance = fixture.tokens[TokenIndex::USDC as usize].balance(&frodo);
-    let frodo_xlm_balance = fixture.tokens[TokenIndex::XLM as usize].balance(&frodo);
+    let frodo_usdc_balance = fixture.tokens[TokenIndex::USDC].balance(&frodo);
+    let frodo_xlm_balance = fixture.tokens[TokenIndex::XLM].balance(&frodo);
     let quote = pool_fixture.pool.fill_auction(&frodo, &0, &samwise);
     assert_approx_eq_abs(
         pool_fixture.reserves[1].b_token.balance(&frodo) - frodo_xlm_btoken_balance,
@@ -239,12 +238,12 @@ fn test_liquidations() {
         1000,
     );
     assert_approx_eq_abs(
-        frodo_usdc_balance - fixture.tokens[TokenIndex::USDC as usize].balance(&frodo),
+        frodo_usdc_balance - fixture.tokens[TokenIndex::USDC].balance(&frodo),
         5981_750792,
         10i128.pow(6),
     );
     assert_approx_eq_abs(
-        frodo_xlm_balance - fixture.tokens[TokenIndex::XLM as usize].balance(&frodo),
+        frodo_xlm_balance - fixture.tokens[TokenIndex::XLM].balance(&frodo),
         14422_6728800,
         SCALAR_7,
     );
@@ -294,11 +293,11 @@ fn test_liquidations() {
     // allow 150 blocks to pass
     fixture.jump(151 * 5);
     // fill bad debt auction
-    let frodo_usdc_pre_fill = fixture.tokens[TokenIndex::USDC as usize].balance(&frodo);
-    let frodo_xlm_pre_fill = fixture.tokens[TokenIndex::XLM as usize].balance(&frodo);
-    let frodo_bstop_pre_fill = fixture.tokens[TokenIndex::BSTOP as usize].balance(&frodo);
+    let frodo_usdc_pre_fill = fixture.tokens[TokenIndex::USDC].balance(&frodo);
+    let frodo_xlm_pre_fill = fixture.tokens[TokenIndex::XLM].balance(&frodo);
+    let frodo_bstop_pre_fill = fixture.tokens[TokenIndex::BSTOP].balance(&frodo);
     let backstop_bstop_pre_fill =
-        fixture.tokens[TokenIndex::BSTOP as usize].balance(&fixture.backstop.address);
+        fixture.tokens[TokenIndex::BSTOP].balance(&fixture.backstop.address);
     let bad_debt_auction_quote =
         pool_fixture
             .pool
@@ -317,7 +316,7 @@ fn test_liquidations() {
         0,
     );
     assert_approx_eq_abs(
-        frodo_usdc_pre_fill - fixture.tokens[TokenIndex::USDC as usize].balance(&frodo),
+        frodo_usdc_pre_fill - fixture.tokens[TokenIndex::USDC].balance(&frodo),
         1993_916931,
         10i128.pow(6),
     );
@@ -332,25 +331,25 @@ fn test_liquidations() {
         0,
     );
     assert_approx_eq_abs(
-        frodo_xlm_pre_fill - fixture.tokens[TokenIndex::XLM as usize].balance(&frodo),
+        frodo_xlm_pre_fill - fixture.tokens[TokenIndex::XLM].balance(&frodo),
         4807_5576270,
         SCALAR_7,
     );
     assert_approx_eq_abs(bad_debt_auction_quote_lot, 5196_8126560, SCALAR_7);
     assert_approx_eq_abs(
-        fixture.tokens[TokenIndex::BSTOP as usize].balance(&frodo) - frodo_bstop_pre_fill,
+        fixture.tokens[TokenIndex::BSTOP].balance(&frodo) - frodo_bstop_pre_fill,
         5196_8126560,
         SCALAR_7,
     );
     assert_approx_eq_abs(
         backstop_bstop_pre_fill
-            - fixture.tokens[TokenIndex::BSTOP as usize].balance(&fixture.backstop.address),
+            - fixture.tokens[TokenIndex::BSTOP].balance(&fixture.backstop.address),
         5196_8126560,
         SCALAR_7,
     );
     //check that frodo was correctly slashed
     let original_deposit = 2_000_000 * SCALAR_7;
-    let pre_withdraw_frodo_bstp = fixture.tokens[TokenIndex::BSTOP as usize].balance(&frodo);
+    let pre_withdraw_frodo_bstp = fixture.tokens[TokenIndex::BSTOP].balance(&frodo);
     fixture
         .backstop
         .queue_withdrawal(&frodo, &pool_fixture.pool.address, &original_deposit);
@@ -360,7 +359,7 @@ fn test_liquidations() {
         .backstop
         .withdraw(&frodo, &pool_fixture.pool.address, &original_deposit);
     assert_approx_eq_abs(
-        fixture.tokens[TokenIndex::BSTOP as usize].balance(&frodo),
+        fixture.tokens[TokenIndex::BSTOP].balance(&frodo),
         pre_withdraw_frodo_bstp + original_deposit - 5196_8126560,
         SCALAR_7,
     );
